requirements.txt
streamlit==1.25.0
pillow
pandas
numpy
plotly
matplotlib
seaborn
